% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cos_similarity_based_extraction.R
\name{Cos_sim_based_extr}
\alias{Cos_sim_based_extr}
\title{detect features that are suspicious for being artefacts produced by column bleed solvent etc.}
\usage{
Cos_sim_based_extr(
  align_MS,
  MS_int_cutoff = 50,
  Cos_cutoff = 0.8,
  Transitivity_cutoff = 0.9,
  Group_size = 3
)
}
\arguments{
\item{align_MS}{matrix where in rows are detected peaks and in columns align individual masses value is mass intensity}

\item{MS_int_cutoff}{mass intensity threshold every intensity below this value will be set to zero}

\item{Cos_cutoff}{threshold for cosine similarity between two mass spectra everything below is set to zero}

\item{Transitivity_cutoff}{transitivity threshold everything below is set to zero}

\item{Group_size}{minimum number of detected peaks in group detected in network any group with members less then this number will be drop}
}
\value{
return table where rows are individual detected peaks second column is number of its group and third column marks if keep the peak or not
}
\description{
This functions uses a fact that spectra of natural occurring compounds are usually
far less similar then spectra of common impurities. It searches for groups that are highly similar inside and
but not as similar between groups. If the group fits parameters it is set as suspicious.
}
\details{
Cos_sim_based_extr
}
\examples{
data("Tab_align_ms")
Cos_res <- Cos_sim_based_extr(Tab_align_ms, Cos_cutoff = 0.8)
summary(Cos_res)



}
